Separates resource request from execution.
METHOD_NUM_CORES and METHOD_NUM_NODES will default to ARCH_NUM_CORES and ARCH_NUM_NODES, but there are times when you want to request more resources and only use pieces of them for each job. This, combined with some changes in the amazon code that generates the header file, facilitates that ability.